Create 8. JSON Exercise Store Current Data Instance
Now using the session storage we want to save this JSON object in a browser. If session storage have data already then it will be back. DataJSON is defined globally but as a empty string then its defined inside a function.
buildCheckboxes called in a window.onload function.
addEvents called inside a function buildCheckboxes
updateJSON is called inside a for loop.